requirements.txt
Flask
openai
python-dotenv
google-cloud-documentai
jinja2
